size_t instead of int for non-negative values

+++ include/maze_parse.h
@@ -1,12 +1,13 @@
 #ifndef MAZE_PARSE_H
 #define MAZE_PARSE_H
 
+#include <stddef.h>
 #define MAX_WIDTH 100
 #define MAX_HEIGHT 100
 
 typedef struct Vertex {
-    int x;
-    int y;
+    size_t x;
+    size_t y;
     int is_wall;
     int visited;
     struct Vertex *parent;
@@ -14,8 +15,8 @@ typedef struct Vertex {
 
 typedef struct {
     Vertex grid[MAX_HEIGHT][MAX_WIDTH];
-    int width;
-    int height;
+    size_t width;
+    size_t height;
 } Maze;
 
 int load_maze(const char *filename, Maze *maze);
